backend: fix copilot hallucinations
What was he cooking

+++ backend/src/model/waypoint.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, ForeignKey, Integer
+from sqlalchemy import Column, ForeignKey, Integer, Float
 from src.db import Base
 
 
@@ -6,5 +6,5 @@ class WaypointModel(Base):
     __tablename__ = "waypoints"
     id = Column(Integer, primary_key=True)
     route_id = Column(Integer, ForeignKey("routes.id"), nullable=False)
-    lat = Column(Integer, nullable=False)
-    lon = Column(Integer, nullable=False)
+    lat = Column(Float, nullable=False)
+    lon = Column(Float, nullable=False)
